Fixes the field property lookup
Since the interface change to set/getDefinedType() I am just not
getting around to making sure that change gets everywhere.  Fix that
for the setProperties() method in the PropertyLook object as well as
adjusting some of the unit tests.

Also needed to remove the checks for primary key at the field level.
I just don't believe that it would have been used.  Something like that
could go back in at a later time if a genuine need arises.  Trying to
keep things as simple as possible here.