Removes the getPhpType() requirement from the Field interface
Adjusted the unit tests to no longer look for that.  Instead, going to
have the field objects create the correct PHP type as a pass through
value cleaner.

Removed the Field\PostgreSQL class.  No real need for it at this point.
There may be a need for something there later, but it sure won't look
like what was there.