A new interface to describe how to clean up data in and out of fields
I had considered not doing anything with values of data in this
library, but after further consideration this is a logical place to
do it.  Every field is already assigned a Field object, and that
object should already have some clue about the boundaries of what is
allowed back into the database.

Some items coming out of the DB should be mostly left alone, like
strings and numbers.  Others need some minimal conversion, like
boolean values.  More complex fields like JSON, Composite, and Arrays
will need to be converted into objects that are more suitable for use
in PHP, then back to the format needed to be stored in the DB.

I suspect that normal usage will have a bit of coercion when
filtering, like truncating a string down to the stated size.  This
will be the default behavior.  I also want to provide a strict mode
that will throw an exception when boundaries are respected.