Shuffling interface responsibilities and simplifying the Field objects
Now that there's a strict mode setting trait, remove all the code
from the field objects that dealt with that.

There is no longer a FieldValues interface, as that is now combined
with the Field interface.

Traits were moved up a directory so as to be a shared resource, so
namespaces were adjusted accordingly.